los componentes peso mortalidad y gasto arreglados a nivel de front y back y falta hacer que al actualizar no se regrese al la pagina contenido admin y se mantenga en el boton seleccionado

+++ frontend/src/api/mortalidad.api.js
@@ -7,31 +7,52 @@ async function getAllMortalidad(){
 
 async function postMortalidad (mortalidadData){
     try {
-        const respuesta=await axios.post(`http://localhost:4000/mortalidad`, mortalidadData); 
-        // console.log("Se agrego una nueva mortalidad",respuesta.data);
-        return respuesta.data;    
+        const respuesta=await axios.post('http://localhost:4000/mortalidad',mortalidadData);
+        if (respuesta.status === 201) {
+            if (respuesta.data !== null) {
+                console.log("informacion de mortalidad agregado",respuesta.data );
+            // }else{
+            //     console.log("respuesta del servidor es null", respuesta.data);
+            }
+            return respuesta.data;
+        } else {
+            return null;
+        }
     } catch (error) {
         console.log("error al guardar mortalidad" , error)
+        throw error;
     }
-    
-}
-async function patchMortalidad (id, mortalidadData) {
-    const respuesta = await axios.put(`http://localhost:4000/mortalidad/${id}`, mortalidadData);
-    return respuesta.data;
 }
 
-async function deleteMortalidad (id){
-    const respuesaEliminar = await axios.delete(`http://localhost:4000/mortalidad/${id}`);
-    if(respuesaEliminar.status==200) {
-        alert ("Se elimino correctamente")
-    } else{
-        alert("No se pudo eliminar el alimento")
+async function patchMortalidad (id, mortalidadData) {
+    try {
+        const respuesta = await axios.put(`http://localhost:4000/mortalidad/${id}`, mortalidadData);
+        return respuesta.data;
+    } catch (error){
+        console.log(error);
+        return false;
     }
+}
+
+async function  eliminarMortalidad(id){
+    try {
+        const mortalidadExiste = await axios.get(`http://localhost:4000/mortalidad/${id}`)
+        if (!mortalidadExiste.data){
+            console.log("no existe la mortalidad a eliminar");
+            return false;
+        }
+        const respuestaEliminar= await axios.delete(`http://localhost:4000/mortalidad/${id}`);
+        return respuestaEliminar.data;
+    } catch (error) {
+        console.log("Error al eliminar la mortalidad", error);
+        throw error;
+   }
 }
 
 export{
     getAllMortalidad,   
     postMortalidad,
     patchMortalidad,
-    deleteMortalidad
+    eliminarMortalidad
+    
 }
